move geoIP structs from cfg.go to collections.go file
